requirements.txt
aiohttp==3.7.4.post0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
backoff==1.11.1
beautifulsoup4==4.10.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
datasets==1.12.1
dill==0.3.4
Django==3.2.7
django-appconf==1.0.5
django-bootstrap4==3.0.1
django-imagekit==4.0.2
emoji==1.5.0
filelock==3.3.0
fsspec==2021.9.0
huggingface-hub==0.0.19
idna==3.2
joblib==1.1.0
multidict==5.1.0
multiprocess==0.70.12.2
numpy==1.21.2
packaging==21.0
pandas==1.3.3
pilkit==2.0
Pillow==8.3.2
pyarrow==5.0.0
pyparsing==2.4.7
pysentimiento==0.2.5
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
regex==2021.9.30
requests==2.26.0
sacremoses==0.0.46
scikit-learn==1.0
scipy==1.7.1
six==1.16.0
sklearn==0.0
soupsieve==2.2.1
spotify-client==1.9.1
spotipy==2.19.0
sqlparse==0.4.2
threadpoolctl==3.0.0
tokenizers==0.10.3
torch==1.9.1
tqdm==4.62.3
transformers==4.11.3
typing-extensions==3.10.0.2
urllib3==1.26.7
wincertstore==0.2
xxhash==2.0.2
yarl==1.6.3
